音を一度に全部読み込む関数作って置き換え

+++ Application/Base/ResultScene.h
@@ -7,7 +7,7 @@ class ResultScene :
     public SceneState
 {
 public:
-    uint32_t png_result_ = 0;       //リザルト画像
+    uint64_t png_result_ = 0;       //リザルト画像
     uint32_t result_BGM_ = 0;
     uint32_t sceneChange_SE_ = 0;   //シーン遷移音
